# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Kaay entreprendre",
        "description": "Cet api va nous permettre de faire communiquer la partie front et la partie back de notre plateforme",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/etude-cas/create": {
            "post": {
                "tags": [
                    "EtudeCas"
                ],
                "summary": "Ajouter une nouvelle étude de cas",
                "description": "Ajoute une nouvelle étude de cas avec les détails fournis",
                "operationId": "createEtudeCas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du secteur",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "secteur_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Etude de cas ajoutée avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/update_etudeCas": {
            "post": {
                "tags": [
                    "EtudeCas"
                ],
                "summary": "Modifier une étude de cas existante",
                "description": "Modifie les détails d'une étude de cas existante avec les données fournies",
                "operationId": "updateEtudeCas",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "contenu": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "secteur_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Etude de cas modifiée avec succès"
                    }
                }
            }
        },
        "/api/etude-cas/archive/{id}": {
            "post": {
                "tags": [
                    "EtudeCas"
                ],
                "summary": "Archiver une étude de cas",
                "description": "Archive une étude de cas spécifique",
                "operationId": "archiveEtudeCas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étude de cas",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Etude de cas archivée avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/api/etude-cas/delete/{id}": {
            "post": {
                "tags": [
                    "EtudeCas"
                ],
                "summary": "Supprimer une étude de cas",
                "description": "Supprime une étude de cas spécifique",
                "operationId": "deleteEtudeCas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'étude de cas",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Etude de cas supprimée avec succès"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
...
}